Add API ability to filter based on unread status or list.
The query parameter "unread=" can now be used to get only the bookmarks
that are either read or unread. If "unread=0", then only the read
bookmarks will be returned. If "unread=1", then only the unread
bookmarks will be returned.

The query parameter "list=" can now be used to get only the bookmarks
that are part of a specific list.