Add Docker support and modify settings.py.
The Dockerfile is used to build the onstop4/bookmarker Docker image.
The docker-compose.yml can be used to provision an environment with
Bookmarker and Postgres running. The migrate script can be used to
perform database migrations on the Postgres database created through
Docker Compose.

Webpack will now generate and store its files in the "static"
directory.

The ALLOWED_HOST environment variable allows you to specify the
host/domain name for the Django site. The previous behavior was to
accept all connections if DEBUG is False.

If DEBUG is True, Bookmarker will use Django's console backend instead
of an SMTP server, so confirmation emails will just be printed to the
console.